Fixup compilation on gcc 4.x/linux 64 systems.
Aliasing between C types (uint8_t et al) and Open ZWave types (uint8 et al) was working
with some compilers/build environments but not others.

To fix this, a new include file zwave_types.h is introduced which defines these types for the purposes of the CGO compile and the Go implementation is redefined in terms of those types.

Thanks to Troy (aka SurfinaNinjas on GitHub) for reporting the issue (#1) and identifying the fix.

Signed-off-by: Jon Seymour <jon@ninjablocks.com>