fix issue revealed by tightening of c-go rules
In go1.6, cgo rules regarding the passing of Go pointers to C libraries
changed to prevent unsafe interactions with the Go garbage collector.

We need to change so that we pass an indirect reference to the Go object
to the c-runtime.

This code has been tested to the extent that is required to address the
particular panic reported with https://github.com/ninjasphere/go-openzwave/issues/4

However, a more extensive regression test has not yet been performed.

Signed-off-by: Jon Seymour <jon@ninjablocks.com>